moved blogCard component inside blog page
