+++ pages/blog/components/blogCard.js
@@ -1,5 +1,5 @@
-import { Element } from '../../js/components/element.js';
-import { ActionButton } from '../../js/components/actionButton.js';
+import { Element } from '../../../js/components/element.js';
+import { ActionButton } from '../../../js/components/actionButton.js';
 
 class BlogCard extends Element {
 	constructor(title, description, year, month, day, fileLinks, order) {
